requirements.txt
fastapi
uvicorn
websockets
deepgram-sdk
anthropic
supabase
sentence-transformers
twilio
httpx
python-dotenv
